+++ platform/windows/windows_terminal_logger.cpp
@@ -5,8 +5,8 @@
 /*                           GODOT ENGINE                                */
 /*                      https://godotengine.org                          */
 /*************************************************************************/
-/* Copyright (c) 2007-2018 Juan Linietsky, Ariel Manzur.                 */
-/* Copyright (c) 2014-2018 Godot Engine contributors (cf. AUTHORS.md)    */
+/* Copyright (c) 2007-2020 Juan Linietsky, Ariel Manzur.                 */
+/* Copyright (c) 2014-2020 Godot Engine contributors (cf. AUTHORS.md).   */
 /*                                                                       */
 /* Permission is hereby granted, free of charge, to any person obtaining */
 /* a copy of this software and associated documentation files (the       */
@@ -45,7 +45,7 @@ void WindowsTerminalLogger::logv(const char *p_format, va_list p_list, bool p_er
 	int len = vsnprintf(buf, BUFFER_SIZE, p_format, p_list);
 	if (len <= 0)
 		return;
-	if (len >= BUFFER_SIZE)
+	if ((unsigned int)len >= BUFFER_SIZE)
 		len = BUFFER_SIZE; // Output is too big, will be truncated
 	buf[len] = 0;
